Update Content-Type and authToken in API request
Modified the `Content-Type` header to `application/json` in the `ExamApp.API.http` file. Replaced the specific `@authToken` value with a placeholder (`xxx`) to indicate that it should be manually assigned during execution. These changes are related to the registration process for the API endpoint.

Update API to use HTTPS and enhance endpoints

- Changed API host from HTTP to HTTPS for secure communication.
- Added and modified endpoints for authentication, user management, exam management, question management, exam results, and answers.
- Specified content type as `application/json` for relevant requests.
- Defined JSON payload structures for various API requests.